Separate firebase file
