+++ src/App.js
@@ -10,28 +10,8 @@ import { ThemeContext } from "./contexts/ThemeContext";
 import { Main, BlogPage, ProjectPage } from "./pages";
 import { BackToTop } from "./components";
 import ScrollToTop from "./utils/ScrollToTop";
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-import { getFirestore } from "firebase/firestore";
 
 import "./App.css";
-
-const firebaseConfig = {
-  apiKey: "AIzaSyADnUc4Nb6zu3P1degBPwZZdvIJKEvzyxw",
-  authDomain: "my-porfolio-c5ab3.firebaseapp.com",
-  projectId: "my-porfolio-c5ab3",
-  storageBucket: "my-porfolio-c5ab3.appspot.com",
-  messagingSenderId: "163404663166",
-  appId: "1:163404663166:web:44f886ade6e26776a34b2f",
-  measurementId: "G-KSZ7ZJ7RW9",
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-const db = getFirestore(app);
-
-export { db };
 
 function App() {
   const { theme } = useContext(ThemeContext);
